move model code to transformer file
